Compiles with go1 (barely tested)
