+++ example/basic.go
@@ -1,16 +1,16 @@
 package main
 
-import "lua51"
+import "golua/lua"
 import "fmt"
 
-func test(L *lua51.State) int {
+func test(L *lua.State) int {
 	fmt.Println("hello world! from go!");
 	return 0;
 }
 
-func test2(L *lua51.State) int {
-	arg := lua51.CheckInteger(L,-1);
-	argfrombottom := lua51.CheckInteger(L,1);
+func test2(L *lua.State) int {
+	arg := lua.CheckInteger(L,-1);
+	argfrombottom := lua.CheckInteger(L,1);
 	fmt.Print("test2 arg: ");
 	fmt.Println(arg);
 	fmt.Print("from bottom: ");
@@ -19,12 +19,12 @@ func test2(L *lua51.State) int {
 }
 
 func main() {
-	var L *lua51.State;
+	var L *lua.State;
 
-	L = lua51.NewState();
+	L = lua.NewState();
 	L.OpenLibs();
 
-	L.GetField(lua51.LUA_GLOBALSINDEX, "print");
+	L.GetField(lua.LUA_GLOBALSINDEX, "print");
 	L.PushString("Hello World!");
 	L.Call(1,0);
