+++ lua/golua.h
@@ -1,3 +1,4 @@
+#include <stdint.h>
 
 typedef struct { void *t; void *v; } GoInterface;
 
@@ -7,7 +8,7 @@ void clua_initstate(lua_State* L);
 unsigned int clua_togofunction(lua_State* L, int index);
 void clua_pushgofunction(lua_State* L, unsigned int fid);
 void clua_pushlightinteger(lua_State* L, int n);
-int clua_tolightinteger(lua_State* L, int index);
+uintptr_t clua_tolightinteger(lua_State* L, int index);
 void clua_setgostate(lua_State* L, GoInterface gostate);
 GoInterface* clua_getgostate(lua_State* L);
 GoInterface clua_atpanic(lua_State* L, unsigned int panicf_id);
