+++ lua/lua.go
@@ -1,30 +1,23 @@
-package lua51
+package lua
 
-//#include <lua.h>
-//#include "golua.h"
-//#include <stdlib.h>
+/*
+#cgo pkg-config: lua5.1
+
+#include <lua.h>
+#include <stdlib.h>
+
+#include "golua.h"
+
+*/
 import "C"
 
 import "unsafe"
 
-
-
-
 //like lua_Writer, but as p will contain capacity, not needed as separate param
 type Writer func(L *State, p []byte, ud interface{});
-//like lua reader, but the return slice has the size, so does 
+//like lua reader, but the return slice has the size, so does
 // we do not need it to be an out param
 type Reader func(L *State, data interface{}) []byte;
-
-//wrapper to keep cgo from complaining about incomplete ptr type
-//export State
-type State struct {
-	s *C.lua_State;
-	//funcs []GoFunction;
-	registry []interface{};
-	//freelist for funcs indices, to allow for freeing
-	freeIndices []uint;
-}
 
 func newState(L *C.lua_State) *State {
 	var newstatei interface{}
@@ -87,40 +80,6 @@ func (L *State) unregister(fid uint) {
 	}
 }
 
-type GoFunction func(*State) int;
-
-//export golua_callgofunction
-func golua_callgofunction(L interface{}, fid uint) int {
-	L1 := L.(*State);
-	f := L1.registry[fid].(GoFunction);
-	return f(L1);
-}
-
-//export golua_gchook
-func golua_gchook(L interface{}, id uint) int {
-	L1 := L.(*State);
-	L1.unregister(id);
-	return 0;
-}
-
-//export golua_callpanicfunction
-func golua_callpanicfunction(L interface{}, id uint) int {
-	L1 := L.(*State);
-	f := L1.registry[id].(GoFunction);
-	return f(L1);
-}
-
-//export golua_idtointerface
-func golua_idtointerface(id uint) interface{} {
-	return id;
-}
-
-//export golua_cfunctiontointerface
-func golua_cfunctiontointerface(f *uintptr) interface{} {
-	return f;
-}
-
-
 func (L *State) PushGoFunction(f GoFunction) {
 	fid := L.register(f);
 	C.clua_pushgofunction(L.s,C.uint(fid));
@@ -140,7 +99,7 @@ func (L *State) PushLightUserdata(ud *interface{}) {
 }
 /*
 //TODO:
-//push pointer as full userdata - mem is go owned, but we 
+//push pointer as full userdata - mem is go owned, but we
 //make a guarantee that lifetime will outlast lua lifetime
 func PushUserdata(L *State, ud interface{}) {
 
@@ -150,14 +109,6 @@ func PushUserdata(L *State, ud interface{}) {
 //old style
 func (L *State) NewUserdata(size uintptr) unsafe.Pointer {
 	return unsafe.Pointer(C.lua_newuserdata(L.s, C.size_t(size)));
-}
-
-
-type Alloc func(ptr unsafe.Pointer, osize uint, nsize uint) unsafe.Pointer;
-//export golua_callallocf
-func golua_callallocf(fp uintptr,	ptr uintptr,
-			    osize uint,			nsize uint) uintptr {
-	return uintptr((*((*Alloc)(unsafe.Pointer(fp))))(unsafe.Pointer(ptr),osize,nsize));
 }
 
 func (L *State) AtPanic(panicf GoFunction) (oldpanicf GoFunction) {
@@ -178,7 +129,7 @@ func (L *State) AtPanic(panicf GoFunction) (oldpanicf GoFunction) {
 		}
 	}
 	//generally we only get here if the panicf got set to something like nil
-	//potentially dangerous because we may silently fail 
+	//potentially dangerous because we may silently fail
 	return nil;
 }
 
@@ -208,7 +159,7 @@ func (L *State) GoPCall(fun GoFunction, ud interface{}) int {
 	return 0;
 }
 
-//TODO: data be a slice? 
+//TODO: data be a slice?
 func (L *State) Dump( writer Writer, data interface{}) int {
 	//TODO:
 	return 0;
@@ -247,7 +198,7 @@ func (L *State) IsBoolean(index int) bool {
 }
 
 func (L *State) IsGoFunction(index int) bool {
-	//TODO:go function is now a userdatum, not a c function, so this will not work 
+	//TODO:go function is now a userdatum, not a c function, so this will not work
 	return C.lua_iscfunction(L.s, C.int(index)) == 1
 }
