--- luaver.sh
@@ -1,9 +1,0 @@
-#!/bin/sh
-
-LUAVER=$(pkg-config --print-provides lua | cut -d ' ' -f1)
-
-if test "$LUAVER" = ""; then
-    LUAVER=lua5.1
-fi
-
-echo $LUAVER
